Use list_images and list_sizes for GCE.
- Move methods to Libcloud base class.
- Fix EC2 and GCE unit tests.
