+++ ipa/ipa_ec2.py
@@ -145,21 +145,6 @@ class EC2Provider(LibcloudProvider):
             entry
         )
 
-    def _get_image(self):
-        """Retrieve NodeImage given the image id."""
-        try:
-            image = self.compute_driver.list_images(
-                ex_image_ids=[self.image_id]
-            )[0]
-        except (IndexError, BaseHTTPError):
-            raise EC2ProviderException(
-                'Image with ID: {image_id} not found.'.format(
-                    image_id=self.image_id
-                )
-            )
-
-        return image
-
     def _get_instance(self):
         """Retrieve instance matching instance_id."""
         try:
@@ -174,22 +159,6 @@ class EC2Provider(LibcloudProvider):
             )
         return instance
 
-    def _get_instance_size(self):
-        """Retrieve NodeSize given the instance type."""
-        instance_type = self.instance_type or EC2_DEFAULT_TYPE
-
-        try:
-            sizes = self.compute_driver.list_sizes()
-            size = [size for size in sizes if size.id == instance_type][0]
-        except IndexError:
-            raise EC2ProviderException(
-                'Instance type: {instance_type} not found.'.format(
-                    instance_type=instance_type
-                )
-            )
-
-        return size
-
     def _launch_instance(self):
         """Launch an instance of the given image."""
         if not self.ssh_key_name:
@@ -199,7 +168,7 @@ class EC2Provider(LibcloudProvider):
 
         instance = self.compute_driver.create_node(
             name=ipa_utils.generate_instance_name('ec2-ipa-test'),
-            size=self._get_instance_size(),
+            size=self._get_instance_size(EC2_DEFAULT_TYPE),
             image=self._get_image(),
             ex_keyname=self.ssh_key_name
         )
